# docs/api/openapi/v1/all.json
{
  "components": {
    "schemas": {
...
      "RestoreRequest": {
        "properties": {
          "confirmPhrase": {
            "type": "string"
          }
        },
        "required": [
          "confirmPhrase"
        ],
        "type": "object"
      },
...
      "UpdateAuditRetentionRequest": {
        "properties": {
          "cron": {
            "type": "string"
          },
          "retentionYears": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "cron",
          "retentionYears"
        ],
        "type": "object"
      },
...
      "UpdateScheduleRequest": {
        "properties": {
          "cron": {
            "type": "string"
          },
          "retention": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "cron",
          "retention"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Aggregate of all Balsm modules \u2014 generated. Do not hand-edit.",
    "title": "Balsm API (aggregate)",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/admin/audit/archives": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 20,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "GET /api/v1/admin/audit/archives \u2014 paged list of audit archives",
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/v1/admin/audit/logs": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 50,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "module",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "GET /api/v1/admin/audit/logs \u2014 paged + filtered audit log",
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/v1/admin/audit/retention": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "GET /api/v1/admin/audit/retention",
        "tags": [
          "Audit"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuditRetentionRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuditRetentionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuditRetentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "PUT /api/v1/admin/audit/retention",
        "tags": [
          "Audit"
        ]
      }
    },
...
    "/api/v1/admin/backups": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 20,
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "GET /api/v1/admin/backups \u2014 paged list of backups",
        "tags": [
          "Backups"
        ]
      },
      "post": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "POST /api/v1/admin/backups \u2014 trigger on-demand backup",
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/v1/admin/backups/schedule": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "GET /api/v1/admin/backups/schedule",
        "tags": [
          "Backups"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "PUT /api/v1/admin/backups/schedule",
        "tags": [
          "Backups"
        ]
      }
    },
    "/api/v1/admin/backups/{id}/restore": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "POST /api/v1/admin/backups/{id}/restore",
        "tags": [
          "Backups"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5550/"
    }
  ],
...
}